fix: simplified use of info-main element
