--- src/view/info-main-view.js
@@ -1,16 +1,0 @@
-import AbstractView from '../framework/view/abstract-view.js';
-
-// сюда по идее должны подтягиваться дестинейшны по порядку с путей назначения (+проверка на уникальность)
-function createTripInfoMainTemplate() {
-  return `<div class="trip-info__main">
-            <h1 class="trip-info__title">Amsterdam &mdash; Chamonix &mdash; Geneva</h1>
-
-            <p class="trip-info__dates">Mar 18&nbsp;&mdash;&nbsp;20</p>
-          </div>`;
-}
-
-export default class TripInfoMainView extends AbstractView {
-  get template() {
-    return createTripInfoMainTemplate();
-  }
-}
