+++ src/view/info-view.js
@@ -1,7 +1,25 @@
 import AbstractView from '../framework/view/abstract-view.js';
 
+function createTripInfoMainTemplate() {
+  return `<div class="trip-info__main">
+            <h1 class="trip-info__title">Amsterdam &mdash; Chamonix &mdash; Geneva</h1>
+
+            <p class="trip-info__dates">Mar 18&nbsp;&mdash;&nbsp;20</p>
+          </div>`;
+}
+
+function createTripInfoCostTemplate() {
+  return `<p class="trip-info__cost">
+            Total: &euro;&nbsp;<span class="trip-info__cost-value">1230</span>
+          </p>`;
+}
+
 function createTripInfoTemplate() {
-  return '<section class="trip-main__trip-info  trip-info"></section>';
+  return /*HTML*/ `
+    <section class="trip-main__trip-info  trip-info">
+      ${createTripInfoMainTemplate()}
+      ${createTripInfoCostTemplate()}
+    </section>`;
 }
 
 export default class TripInfoView extends AbstractView {
